feat(tui): add the Laravel Doctor as an on-demand site tab
The web dashboard gained a Laravel Doctor panel that runs app-level health checks, but the TUI had no equivalent, so diagnosing a site from the terminal meant dropping to the CLI. The site detail pane now carries a fifth Doctor tab for Laravel sites, opened with 5, showing the same checks the web runs: APP_KEY, .env drift against .env.example (warning only on keys the code reads without a default), the APP_DEBUG-in-production footgun, the public/storage symlink, and pending migrations. Because the migrations check execs artisan in the container the run is on-demand rather than on every render: pressing 5 kicks it off and pressing it again re-runs, with the result cached per site so tabbing away and back is instant. The tab is read-only and names the suggested fix instead of running it, keeping destructive commands like migrate on the CLI.

To avoid a second copy of the check logic the doctor moves into a new internal/sitedoctor package that both the web handler and the TUI call, with the web route now a thin delegator over it.
