+++ internal/tui/doctor.go
@@ -1,0 +1,147 @@
+package tui
+
+import (
+	"context"
+	"fmt"
+	"strings"
+
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/geodro/lerd/internal/sitedoctor"
+	"github.com/geodro/lerd/internal/siteinfo"
+)
+
+// doctorResultMsg carries a finished Laravel Doctor run back into the model.
+// site keys the result to the site it was run for, so a result that lands after
+// the user has moved on to another site is discarded rather than shown against
+// the wrong site.
+type doctorResultMsg struct {
+	site string
+	resp sitedoctor.Response
+}
+
+// doctorCmd runs the Laravel Doctor checks for a site off the main loop. The
+// migrations check execs artisan in the container (up to ~25s), so this must
+// never run inline in Update — it returns a message the handler folds into the
+// model when it completes.
+func doctorCmd(siteName, path string) tea.Cmd {
+	return func() tea.Msg {
+		return doctorResultMsg{site: siteName, resp: sitedoctor.Run(context.Background(), path)}
+	}
+}
+
+// openDoctorTab switches to the Doctor tab for a Laravel site and kicks off a
+// fresh run when there's no cached result for it yet or the user re-pressed the
+// shortcut while already on the tab (an explicit refresh). Returns nil for
+// non-Laravel sites so the shortcut is a no-op there.
+func (m *Model) openDoctorTab() tea.Cmd {
+	s := m.currentSite()
+	if !siteIsLaravel(s) {
+		return nil
+	}
+	refresh := m.siteTab == tabSiteDoctor
+	m.siteTab = tabSiteDoctor
+	m.detailScroll = 0
+	m.focus = paneDetail
+	if m.doctorLoading {
+		return nil
+	}
+	if refresh || m.doctorSite != s.Name {
+		m.doctorLoading = true
+		m.doctorChecks = nil
+		m.doctorSite = s.Name
+		return doctorCmd(s.Name, s.Path)
+	}
+	return nil
+}
+
+// doctorStatusGlyph maps a check status to the shared state glyph + style so the
+// Doctor panel reads like the rest of the TUI: green pass, amber warn, red fail,
+// grey unknown (a check that couldn't run, e.g. the app is down).
+func doctorStatusGlyph(status string) string {
+	switch status {
+	case sitedoctor.StatusOK:
+		return runningStyle.Render(glyphRunning)
+	case sitedoctor.StatusWarn:
+		return suspendedStyle.Render(glyphPaused)
+	case sitedoctor.StatusFail:
+		return failingStyle.Render(glyphFailing)
+	default:
+		return stoppedStyle.Render(glyphStopped)
+	}
+}
+
+func doctorStatusText(status string) string {
+	switch status {
+	case sitedoctor.StatusOK:
+		return runningStyle.Render("ok")
+	case sitedoctor.StatusWarn:
+		return suspendedStyle.Render("warn")
+	case sitedoctor.StatusFail:
+		return failingStyle.Render("fail")
+	default:
+		return dimStyle.Render("unknown")
+	}
+}
+
+// siteDoctorContentLines renders the focused site's Laravel Doctor panel: a
+// running placeholder while the checks are in flight, then one row per check
+// with its status, detail, and suggested fix command. Read-only — the panel
+// names the fix (key:generate, migrate, …) rather than running it, so a stray
+// keypress can't migrate a database from a status view.
+func siteDoctorContentLines(m *Model, site *siteinfo.EnrichedSite, innerW int) []string {
+	out := make([]string, 0, 32)
+	out = append(out, renderSiteTabHeader(tabSiteDoctor, innerW, siteIsLaravel(site))...)
+	add := func(s string) { out = append(out, padToWidth(clipLine(s, innerW), innerW)) }
+
+	if site == nil {
+		add(dimStyle.Render("  no site selected"))
+		return out
+	}
+
+	add(sectionStyle.Render("Laravel Doctor") + dimStyle.Render("  ·  "+site.PrimaryDomain()))
+	add("")
+
+	if m.doctorLoading && m.doctorSite == site.Name {
+		add(dimStyle.Render("  running checks… (migrations may take a moment)"))
+		return out
+	}
+	if m.doctorSite != site.Name {
+		add(dimStyle.Render("  press ") + accentStyle.Render("5") + dimStyle.Render(" to run the checks"))
+		return out
+	}
+	if len(m.doctorChecks) == 0 {
+		add(dimStyle.Render("  no checks ran — not a Laravel site, or nothing applied"))
+		return out
+	}
+
+	failures, warnings := 0, 0
+	for _, c := range m.doctorChecks {
+		switch c.Status {
+		case sitedoctor.StatusFail:
+			failures++
+		case sitedoctor.StatusWarn:
+			warnings++
+		}
+	}
+	switch {
+	case failures > 0 || warnings > 0:
+		add(dimStyle.Render(fmt.Sprintf("  %d failing · %d warning", failures, warnings)))
+	default:
+		add(runningStyle.Render("  all checks pass"))
+	}
+	add("")
+
+	for _, c := range m.doctorChecks {
+		name := strings.ReplaceAll(c.Name, "_", " ")
+		add(fmt.Sprintf("  %s %s  %s", doctorStatusGlyph(c.Status), padRight(name, 14), doctorStatusText(c.Status)))
+		if c.Detail != "" {
+			add(dimStyle.Render("      " + c.Detail))
+		}
+		if c.Fix != "" {
+			add(dimStyle.Render("      fix: ") + accentStyle.Render(c.Fix))
+		}
+	}
+	add("")
+	add(dimStyle.Render("  press ") + accentStyle.Render("5") + dimStyle.Render(" to re-run"))
+	return out
+}
